+++ src/main/java/com/influxdb/v3/client/InfluxDBClient.java
@@ -526,6 +526,7 @@ public interface InfluxDBClient extends AutoCloseable {
      *   <li>database - database (bucket) name</li>
      *   <li>precision - timestamp precision when writing data</li>
      *   <li>gzipThreshold - payload size size for gzipping data</li>
+     *   <li>writeNoSync - skip waiting for WAL persistence on write</li>
      * </ul>
      *
      * @param connectionString connection string
@@ -558,6 +559,7 @@ public interface InfluxDBClient extends AutoCloseable {
      *   <li>INFLUX_DATABASE - database (bucket) name</li>
      *   <li>INFLUX_PRECISION - timestamp precision when writing data</li>
      *   <li>INFLUX_GZIP_THRESHOLD - payload size size for gzipping data</li>
+     *   <li>INFLUX_WRITE_NO_SYNC - skip waiting for WAL persistence on write</li>
      * </ul>
      * Supported system properties:
      * <ul>
@@ -568,6 +570,7 @@ public interface InfluxDBClient extends AutoCloseable {
      *   <li>influx.database - database (bucket) name</li>
      *   <li>influx.precision - timestamp precision when writing data</li>
      *   <li>influx.gzipThreshold - payload size size for gzipping data</li>
+     *   <li>influx.writeNoSync - skip waiting for WAL persistence on write</li>
      * </ul>
      *
      * @return instance of {@link InfluxDBClient}